refactoring createUser and deleting deprecated createdUserAuth, refactor on unit tests
