+++ src/modules/user/services/user.service.ts
@@ -1,9 +1,7 @@
-import { Injectable } from '@nestjs/common';
-import { Transactional } from 'typeorm-transactional-cls-hooked';
+import { BadRequestException, Injectable } from '@nestjs/common';
 
-import { UserCreateDto, UserDto } from '../dtos';
+import { UserDto } from '../dtos';
 import { UserAuthEntity, UserEntity } from '../entities';
-import { UserAuthService } from '../services';
 import {
   generateHash,
   generateRandomInteger,
@@ -14,39 +12,44 @@ import {
 import { PageDto, PageOptionsDto } from '../../../common/dtos';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DataSource, Repository } from 'typeorm';
-import { PinCodeGenerationErrorException } from '../exceptions';
+import {
+  PinCodeGenerationErrorException,
+  UserCreationException,
+} from '../exceptions';
 import { RoleType } from '../constants/role-type.constant';
+import { PostgresErrorCode } from 'src/modules/database/constraints';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(UserEntity)
     private readonly _userRepository: Repository<UserEntity>,
-    private readonly _userAuthService: UserAuthService,
     private DataSourceService: DataSource,
   ) {}
 
   public async createUser(userCreateDto: any): Promise<any> {
-    await this.DataSourceService.manager.transaction(
-      async (transactionalEntityManager) => {
-        const user = await transactionalEntityManager.save(
-          UserEntity,
-          userCreateDto,
-        );
-        const pinCode = await this._createPinCode();
-        const password = await generateHash(user.password);
-        const createdUser = { ...userCreateDto, user, pinCode, password };
-        await transactionalEntityManager.save(UserAuthEntity, createdUser);
+    try {
+      await this.DataSourceService.manager.transaction(
+        async (transactionalEntityManager) => {
+          const user = await transactionalEntityManager.save(
+            UserEntity,
+            userCreateDto,
+          );
+          const pinCode = await this._createPinCode();
+          const password = await generateHash(user.password);
+          const createdUser = { ...userCreateDto, user, pinCode, password };
+          await transactionalEntityManager.save(UserAuthEntity, createdUser);
 
-        return this.findUser({ uuid: user.uuid });
-      },
-    );
+          return this.findUser({ uuid: user.uuid });
+        },
+      );
+    } catch (error) {
+      if (error?.code === PostgresErrorCode.UniqueViolation) {
+        throw new BadRequestException('User with that email already exists');
+      }
 
-    //await this._userRepository.save(user);
-
-    // const createdUser = { ...userCreateDto, user };
-
-    //  await Promise.all([this._userAuthService.createUserAuth(createdUser)]);
+      throw new UserCreationException(error);
+    }
   }
   private async _createPinCode(): Promise<number> {
     const pinCode = this._generatePinCode();
