+++ src/modules/user/tests/services/auth/user-auth-service.spec.ts
@@ -3,7 +3,6 @@ import { getRepositoryToken } from "@nestjs/typeorm";
 import { MockType } from "src/common/utils/mocks";
 import { userServiceMock } from "src/common/utils/mocks/auth-service.mock";
 import { repositoryMockFactory } from "src/common/utils/mocks/repository-factory.mock";
-import { UserCreationException } from "src/modules/user/exceptions";
 import { Repository } from "typeorm";
 import { UserAuthEntity, UserEntity } from "../../../entities";
 import { UserAuthService, UserService } from "../../../services";
@@ -51,17 +50,6 @@ describe('UserAuthServiceMock1', () => {
       expect(userResult).toBeDefined();
     });
 
-    it('should throw UserCreationException when user doesn´t exist', async () => {
 
-      try {
-        const user : Partial<any> = { uuid:'a', email: 'agmi@gmail.com'};
-         await service.createUserAuth(user)
-      } catch (error) {
-        expect(error).toBeInstanceOf(UserCreationException)
-
-      }
-
-
-  });
   });
